requirements.txt
pydantic>=2.5.0
pdfplumber>=0.11.0
PyMuPDF>=1.23.0
PyPDF2>=3.0.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
cryptography>=41.0.0
Pillow>=10.0.0
pytesseract>=0.3.10
# Optional OCR engines (Phase 2)
# easyocr>=1.7.0  # GPU-accelerated OCR
# paddleocr>=2.7.0.3  # Fastest GPU OCR
# pdf2image>=1.16.3  # PDF to image conversion
